+++ container-engine-lib/lib/backend_impls/docker/object_attributes_provider/object_attributes_provider.go
@@ -15,13 +15,11 @@ import (
 )
 
 const (
-	engineServerNamePrefix        = "kurtosis-engine"
-	logsDatabaseServerNamePrefix  = "kurtosis-logs-db"
-	logsCollectorServerNamePrefix = "kurtosis-logs-collector"
+	engineServerNamePrefix  = "kurtosis-engine"
+	logsDatabaseNamePrefix  = "kurtosis-logs-db"
+	logsCollectorNamePrefix = "kurtosis-logs-collector"
 
-	//We always use the same name because we are going to have only one instance of this volume,
-	//so when the engine is restarted it mounts the same volume with the previous logs
-	logsDatabaseVolumeName = logsDatabaseServerNamePrefix + "-vol"
+	logsDatabaseVolumeName = logsDatabaseNamePrefix + "-vol"
 )
 
 type DockerObjectAttributesProvider interface {
@@ -33,13 +31,13 @@ type DockerObjectAttributesProvider interface {
 		grpcProxyPortSpec *port_spec.PortSpec,
 	) (DockerObjectAttributes, error)
 	ForEnclave(enclaveId enclave.EnclaveID) (DockerEnclaveObjectAttributesProvider, error)
-	ForLogsDatabaseServer(
+	ForLogsDatabase(
 		engineGUID engine.EngineGUID,
 		httpApiPortId string,
 		httpApiPortSpec *port_spec.PortSpec,
 	) (DockerObjectAttributes, error)
 	ForLogsDatabaseVolume(engineGUID engine.EngineGUID) (DockerObjectAttributes, error)
-	ForLogsCollectorServer(
+	ForLogsCollector(
 		engineGUID engine.EngineGUID,
 		forwardPortId string,
 		forwardPortSpec *port_spec.PortSpec,
@@ -120,14 +118,14 @@ func (provider *dockerObjectAttributesProviderImpl) ForEnclave(enclaveId enclave
 	return newDockerEnclaveObjectAttributesProviderImpl(enclaveIdLabelValue), nil
 }
 
-func (provider *dockerObjectAttributesProviderImpl) ForLogsDatabaseServer(
+func (provider *dockerObjectAttributesProviderImpl) ForLogsDatabase(
 	engineGUID engine.EngineGUID,
 	httpApiPortId string,
 	httpApiPortSpec *port_spec.PortSpec,
 ) (DockerObjectAttributes, error) {
 	nameStr := strings.Join(
 		[]string{
-			logsDatabaseServerNamePrefix,
+			logsDatabaseNamePrefix,
 			string(engineGUID),
 		},
 		objectNameElementSeparator,
@@ -164,14 +162,14 @@ func (provider *dockerObjectAttributesProviderImpl) ForLogsDatabaseServer(
 	return objectAttributes, nil
 }
 
-func (provider *dockerObjectAttributesProviderImpl) ForLogsCollectorServer(
+func (provider *dockerObjectAttributesProviderImpl) ForLogsCollector(
 	engineGUID engine.EngineGUID,
 	forwardPortId string,
 	forwardPortSpec *port_spec.PortSpec,
 ) (DockerObjectAttributes, error) {
 	nameStr := strings.Join(
 		[]string{
-			logsCollectorServerNamePrefix,
+			logsCollectorNamePrefix,
 			string(engineGUID),
 		},
 		objectNameElementSeparator,
@@ -209,12 +207,10 @@ func (provider *dockerObjectAttributesProviderImpl) ForLogsCollectorServer(
 }
 
 func (provider *dockerObjectAttributesProviderImpl) ForLogsDatabaseVolume(engineGUID engine.EngineGUID) (DockerObjectAttributes, error) {
-	nameStr := logsDatabaseVolumeName
-	name, err := docker_object_name.CreateNewDockerObjectName(nameStr)
+	name, err := docker_object_name.CreateNewDockerObjectName(logsDatabaseVolumeName)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred creating a Docker object name object from string '%v'", nameStr)
+		return nil, stacktrace.Propagate(err, "An error occurred creating a Docker object name object from string '%v'", logsDatabaseVolumeName)
 	}
-
 	engineGuidLabelValue, err := docker_label_value.CreateNewDockerLabelValue(string(engineGUID))
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating the engine GUID Docker label from string '%v'", engineGUID)