+++ container-engine-lib/docs/changelog.md
@@ -3,6 +3,8 @@
 * Added the `ForLogsDatabaseServer` and the `ForLogsCollectorServer` Docker objects attribute
 * Added `LogsCollectorTypeDockerLabelValue`, `ModuleContainerTypeDockerLabelValue` and `LogsDbDataVolumeTypeDockerLabelValue` Docker labels
 * Added `EngineGUIDDockerLabelKey` Docker label key
+* Added `LogsDatabase` interface for defining centralized logs database behaviour
+* Added `Loki` logs database implementation and the Kurtosis configuration values
 
 # 0.36.1
 ### Changes